# docs/api-spec.yaml
openapi: '3.0.3'
info:
  title: GitHub Security Demo API
  description: |
    Sample API for demonstrating GitHub Actions security workflows.
    This API contains intentional vulnerabilities for DAST scanning demos.
  version: '1.0.0'
  contact:
    name: GitHub Sales Engineering

servers:
  - url: http://localhost:3000
    description: Local development

paths:
  /health:
    get:
      summary: Health check
      tags: [System]
      responses:
        '200':
          description: Service is healthy
          content:
            application/json:
              schema:
                type: object
                properties:
                  status:
                    type: string
                  timestamp:
                    type: string
                  version:
                    type: string

  /api/auth/register:
    post:
      summary: Register a new user
      tags: [Authentication]
      requestBody:
        required: true
        content:
          application/json:
            schema:
              type: object
              required: [username, email, password]
              properties:
                username:
                  type: string
                  example: johndoe
                email:
                  type: string
                  format: email
                  example: john@example.com
                password:
                  type: string
                  format: password
                  minLength: 8
                  example: SecureP@ss123
      responses:
        '201':
          description: User registered successfully
        '400':
          description: Missing required fields

  /api/auth/login:
    post:
      summary: Authenticate user
      tags: [Authentication]
      requestBody:
        required: true
        content:
          application/json:
            schema:
              type: object
              required: [username, password]
              properties:
                username:
                  type: string
                password:
                  type: string
                  format: password
      responses:
        '200':
          description: Authentication successful
        '401':
          description: Invalid credentials

  /api/data:
    get:
      summary: Get secure project data
      tags: [Data]
      security:
        - bearerAuth: []
      responses:
        '200':
          description: Project data retrieved
        '401':
          description: Unauthorized

  /api/search:
    get:
      summary: Search projects
      tags: [Data]
      security:
        - bearerAuth: []
      parameters:
        - name: query
          in: query
          required: true
          schema:
            type: string
      responses:
        '200':
          description: Search results

  /api/upload:
    post:
      summary: Upload file
      tags: [Data]
      security:
        - bearerAuth: []
      requestBody:
        content:
          multipart/form-data:
            schema:
              type: object
              properties:
                file:
                  type: string
                  format: binary
      responses:
        '200':
          description: File uploaded

  /api/admin/users:
    get:
      summary: List all users (admin)
      tags: [Admin]
      security:
        - bearerAuth: []
      responses:
        '200':
          description: User list

components:
  securitySchemes:
    bearerAuth:
      type: http
      scheme: bearer
      bearerFormat: JWT